Update: Refactor mce
