+++ src/widgets/tiny-mce/api/index.ts
@@ -1,0 +1,1 @@
+export { generateTOC, setupTOCPlugin } from './tiny-mce-utils';
